Tidy up the project to put it on Github
I am effectively archiving this project. It has taught me a great deal
about how to write interactive simulations and games, and how to scale
them to support many agents.

The next stage would be to write a real game. If I do that I want to
start afresh with tests and profiling.